UI Design

+++ SceneForgeAI/Assets/Core/Editor/Communication/AIResponse.cs
@@ -7,10 +7,10 @@ public class AIResponse
     public AIMessage message { get; set; }
     public string done_reason { get; set; }
     public bool done { get; set; }
-    public int total_duration { get; set; }
-    public int load_duration { get; set; }
+    public ulong total_duration { get; set; }
+    public ulong load_duration { get; set; }
     public int prompt_eval_count { get; set; }
     public int prompt_eval_duration { get; set; }
     public int eval_count { get; set; }
-    public int eval_duration { get; set; }
+    public ulong eval_duration { get; set; }
 }
